refactor: chess 폴더로 전체 묶기 및 폴더명 변경
