+++ java-chess/src/main/java/org/example/chess/Board.java
@@ -1,6 +1,6 @@
-package org.example;
+package org.example.chess;
 
-import org.example.character.Pawn;
+import org.example.chess.pieces.Pawn;
 
 import java.util.ArrayList;
 import java.util.List;
