+++ java-chess/src/main/java/org/example/chess/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package org.example.character;
+package org.example.chess.pieces;
 
 public class Pawn {
     private Color color;
